Make WantsOnChange always true on Button
If a dialog has no "interactive" UIBlock,
it will automatically add an awkward "send interaction" button.
This button always appears on the first row and column,
potentially overlapping with other widgets.
This might lead to confusion if the dialog does have a button (one without a event handler).
So we think it is better to always make a button "interactive" to prevent this.

Change-Id: I656ca6f3e410b4a44a7645d4a1576eef5905f509